Use hacky hash routing
Using hashes to track routes in urls is better for chrome
extensions. This is because the default of providing a route at
/settings for example doesn't map in extension land to a .html

There is probably a way of intercepting route changes in
extensions for urls and redirecting them but this would require
escalated privledges and be harder to test.

The other method of generating a prerendered .html file for every
route almost worked except svelte's routing doesn't recognize
a route at its .html location once it's loaded. It wants to deal
with goto("/settings"), it doesn't recognize
goto("/settings.html").

The hacky workaround I have here works pretty well. Every route
change that happens it goes and replaces the current route with
a corresponding hash representation.

If someone reloads the page the hash representation needs to be
mapped back to its corresponding route and then navigated to
that route by svelte's `goto`. Upon entering that route it will
again have its url replaced with its hash match.

This method also helps encapsulate the swapping of the routes
to be friendly for local dev server or extension. In the case of
the extension the hash is at "index.html#route" and when being
served by the local dev server it is replaced with "/#route".

While this is hacky it is consistent and allows for everything
to be handled entirely "in app". If this app were to increase the
number of routes it had to manage then this hash mapping would
become less sustainable and maybe more research into how other
svelte apps handle this challenge would be warranted.