Add hacky index.html route
This commit adds a route that is as an ugly hack for when
"index.html" loads and it tries to load the corresponding route.
Without this "index.html" route the router can't find a matching
route, throws an error, but still ends up at the default route
(so the user doesn't notice the issue) but it leaves an ugly
error in the console.

By having an "index.html" route which is a shortcoming of being
able to serve from static files like within extensions, this
route kicks things off by navigating to the root route "/".