Move background.ts within src
This moves background.ts to be within the src/chrome folder. This is
a better spot that within the scripts folder as it is now because it
is a part of the src that is built for the app, even if its not
relevant for the svelte front-end part of the app that's built.